formats and updates

+++ Final-Project/uart_helper.h
@@ -1,8 +1,10 @@
 #ifndef UART_HELPER_H_
 #define UART_HELPER_H_
 
-void uart_init(void);
-void uart_send_str(const char *);
-uint8_t uart_get_char(void);
+#include <stdint.h>
+
+void    uart_init(void);
+void    uart_send_str(const char *);
+int16_t uart_get_char(void);
 
 #endif /* UART_HELPER_H_ */
